// functions/src/cloud_functions/bets/models/games/nhl_game.ts
export interface NhlGame {
  GameID?: number;
  Season?: number;
  SeasonType?: number;
  Status?: string;
  Day?: Date;
  DateTime?: Date;
  Updated?: Date;
  IsClosed?: boolean;
  AwayTeam?: string;
  HomeTeam?: string;
  AwayTeamID?: number;
  HomeTeamID?: number;
  StadiumID?: number;
  Channel?: null | string;
  Attendance?: null;
  AwayTeamScore?: number;
  HomeTeamScore?: number;
  Period?: null;
  TimeRemainingMinutes?: null;
  TimeRemainingSeconds?: null;
  AwayTeamMoneyLine?: number;
  HomeTeamMoneyLine?: number;
  PointSpread?: number;
  OverUnder?: number;
  GlobalGameID?: number;
  GlobalAwayTeamID?: number;
  GlobalHomeTeamID?: number;
  PointSpreadAwayTeamMoneyLine?: number;
  PointSpreadHomeTeamMoneyLine?: number;
  LastPlay?: null;
  GameEndDateTime?: Date;
  HomeRotationNumber?: number;
  AwayRotationNumber?: number;
  NeutralVenue?: null;
  OverPayout?: number;
  UnderPayout?: number;
  Periods?: any[];
}