// functions/src/cloud_functions/bets/models/bets/nhl_bet.ts
export interface NhlBet {
  dateTime: Date;
  totalGameScore: number;
  week: string;
  winningTeam: string;
  awayTeam: string;
  stillOpen: boolean;
  betTeam: string;
  clientVersion: string;
  homeTeamCity: string;
  uid: string;
  gameStartDateTime: string;
  odds: number;
  homeTeam: string;
  id: string;
  betPointSpread: number;
  awayTeamCity: string;
  gameId: number;
  league: string;
  winningTeamName: string;
  betType: string;
  betProfit: number;
  betAmount: number;
  isClosed: boolean;
  homeTeamScore: number;
  betOverUnder: number;
  awayTeamName: string;
  awayTeamScore: number;
  dataProvider: string;
  homeTeamName: string;
  username: string;
  status: string;
}